requirements.txt
#本项目所用的依赖包如下:
requests>=2.18.4
Flask>=1.0.2
Flask_restful>=0.3.6
pymongo>=3.7
uniout>=0.3.7
pandas>=0.23.3
xlrd>=1.1.0
bs4>=0.0.1
xmltodict>=0.11.0
PyMySQL>=0.9.2
TestLink-API-Python-client==0.6.4
lxml
openpyxl
